Add files via upload
This information provides an overview of the percentage of people who have received COVID-19 vaccinations. It contains information on how many people have received the vaccine. This information can be utilized to obtain insight into the efficacy of immunization programmers and to identify locations where additional resources may be required. It can also be used to track the progress of the immunization campaign as well as the safety and efficacy of the vaccine.