Add files via upload
Analyzing the connection between movie revenue and ratings is the focus of this study. Data on movie income, audience ratings, genre, and year of release are all included. The initiative will examine the relationship between movie revenue and ratings using statistical techniques. The project's findings will be utilized to learn more about the motion picture business and to assist producers and studios in making wise choices regarding their endeavor's.